[FIX] sap.m.StepInput - key navigation now works fine in table after pressing of F2
* there was a problem with navigation with arrows up/down inside the StepInput,
when it is placed inside a table, keyboard is used to navigate inside it, and F2 key
is pressed (which gives focus to the StepInput inside of a table row). Then, using
of up/down keys increases/decreases the value onece, but then the control loses the
focus.

* Now, inrease/decrease with keys works fine, and control does not lose focus.

Change-Id: Iecd4642f3bbd463540f9c34d45f0799de31d3d40
CR-Id: 002075125800002669002020
BCP: 002075129400000013742020
(cherry picked from commit fd3b087d7d11adcec8450fcebf6f43840c43e6a3)