[FIX] sap.f.ShellBar: Configurating title without menu now possible
Problem:
1.When showMenuButton property of sap.f.ShellBar is set to 'false' we show the primary title
inside MenuButton of MegaMenu control on the left, regardless if the menu has menu items attached
to it. It is confusing, because the button seems clickable, but no click event is associated
with it.
2.Also, when we want to attach menu to the title (showMenuButton === false and sap.m.Menu
aggregation is attached), but we have no Logo image (homeIcon) on mobile we simply show empty
menu button

Solution:
1.We limit the menuButton to only show if there are actual menu to be shown and instead
we show the primary title of the control as text.

2. We attach the Primary title with relative truncation to the menu button if no logo image
is provided.
BCP: 2070227114
Fixes: https://github.com/SAP/openui5/issues/2888

Change-Id: I034379c4d8a0965f71949bf90a150506cdb80682