[INTERNAL] MaskInput: Fix Unit tests for jQuery 3 migration
The tests fail with jQuery 3 because the tests are wrongly written.
The issues are actually unrelated to jQuery 3.

In jQuery 2, a call like 'jQuery(oControl).focus()' accidentally
behaves like 'oControl.focus()'.

jQuery misinterprets oControl as a DOM element and calls focus() on it.
This does not break since the sap.ui.core.Element class implements
a focus() method which is inherited by all Controls.

Most other usages of jQuery(oControl) will fail completely.

Change-Id: I973f794f4d81aa76890257df5d7843aaaebe9d70