[INTERNAL] m.Table: fix gap in sticky when scrolled
There was a padding of 1px applied to the table element when sticky is enabled.
This is done so that the focus outline is visible on all 4 sides.
Due to this padding, in Safari browser, the content of the table was visible
when scrolled behind the sticky elements. Hence to fix the issue, the padding
is changed to margin so that the items are not visible anymore when scrolled
behind the sticky element.
BCP: 2070039988

Change-Id: I3cb92f64d1819aee0d3b09b7f454ad3b6d5f46bf