[FIX] sap.m.PlanningCalendarHeader - View switch remains the same after overflow
* until now, the View switch of the Planning Calendar sometimes changes it's type
from Select to Segmented Button after overflow. This happens when there are more
than 4 views defined for the calendar - in that case Select is used instead of
SegmentedButton. But after the overflow, View switch was displayed as Segmented
Button no matter what it was before the overflow.

* Now if the View switch is Select, after the overflow it remains Select, and if
it is SegmentedButton, it remains SegmentedButton.

Change-Id: I050690fdb6d0aa6487a99b8b77647de476dc2213
BCP: 2040306340