[INTERNAL][FIX] Fixing the creation of app variants on top of app
variants.
- If a user creates app variants based on the running app(variant), then
the selector containing root control of the app(variant) will be used.
If there are dirty changes in the session, they will be taken over by
the app variant.
- If a user creates app variants based on other app(variant) in App
Variant overview dialog, then the selector is a JSON Object which
contains appId and appVersion will be used. If there are any existing
changes for the underlying app(variant) in backend, they will be copied
for the app variant.
- Fixing the inline change creation: If the inline change type does not
expect the texts property, it will be removed by the client before
creation of the change.

Change-Id: I143d00129c7a63f4d8d401b18c536902c56b5b25
BCP: 2030285522