[FEATURE] sap.m.Panel: Active area for expand/collapse enlarged
Until now, the only way to expand/collapse the panel was the small
click area on the icon. With this change the whole header will become
clickable, to make the interaction easier and faster.
Now, when we do not have a headerToolbar provided, the whole header
acts as an active area. Additionaly, we replace the button, with an icon
in this case.

Change-Id: Idfe72f005ce0712daa96499aed95c612deb1eb8e
JIRA: BGSOFUIRILA-2818