[INTERNAL] sap.uxap.ObjectPageLayout: setSelectedSection func optimized
When setSelectedSection method is called with argument of null, the
first visible Section should be selected. This logic is now optimized,
so that some unneeded steps (including the rebuilding of the AnchorBar)
are prevented, in order to improve performance.
The new implementation is based on the fact that, if there is a change
in the Sections/SubSections visibility, the whole layout recalculation
will be done anyway (not depending on the logic of the
setSelectedSection, when called with null).

Change-Id: I140af7ee37bfedda00ce4b93a676cf886bece503
JIRA: BGSOFUIPIRIN-4601