[FIX] sap.m.NavContainer: Ensured pages stack cleared upon removeAllPages
Problem:
When the app removes all pages of a navContainer, the pages are initially removed
from the internal pageStack, but then the first is added back, because of the call to
_ensurePageStackInitialized (which always called before the stack is manipulated)

Solution:
Ensured that the "pages" aggregaton is cleared before manipulating the stack.

Note:
The problem will still appear, if the app did not clear the "initialPage"
association before removing all pages (because the initial page will be added
back after first removing it).
The fix was not extended to cover this case as well, because:
-- Fixing that appears too risky for backward comparibility (as there are various cases,
where the "initialPage" is added to the internal pageStack before being added to the "pages"
aggregation e.g. in getCurrentPage(), so changing this may break existing apps).
-- This case is not yet seen in an actual app. The app that raised the issue,
have their "initialPage" assosiation cleared (or never set) before calling
removeAllPages. (we may ask apps to clear the association or re-assess the
control logic if the case really happens in the future)

Change-Id: I2c233ed398928b0bfaa13f02d31b5bdc120bdce3
Fixes: https://github.com/SAP/openui5/issues/2915
BCP: 2070249903