[INTERNAL][FIX] Demo Kit: Fixed Documentation tab RTL languages layout
Problem:
Buttons are incorrectly aligned in the Documantation Heading in Arabic.

Root Cause:
The container of the buttons overflows out of bounds because it has
its width property set to "100%", and it has paddings at the same time.
This is not an issue in LTR languages, since all the content is on the
left side of the panel.

Solution:
We set the "box-sizing" property of the container of the buttons
to "border box". This way the overflow is omitted since the paddings
are part of the "100%" width.

Change-Id: I22591fac7b537259dc4bd4ae7c0f2db6c40e2795
BCP: 2080138006