[INTERNAL] ListBase: accessibility improvements
The issue reported in the incident below has is already fixed with since
version 1.73. The gerrit change where the accImprovements fix was done is
I5d500cacf311ffaefd1ed82753896b0baabda2ed. This change is only cherry picking
the required fix.
The issue is that when the application adds ariaLabelledBy assocication for
the table, the announcement was not done by JAWS inspite of the aria label
added. This aria label is now moved to the root element so that JAWS can
announce the associated aria labels correctly.
BCP: 2070244288
CR-Id: 002075125900001344982020

Change-Id: Ib51efcc749c83b515fb1872c1911e432564b8c76