[FIX] sap.m.OverflowToolbar: GenericTags are not cut in DynamicPageTitle
Problem: When visibility of GenericTags, placed in OverflowToolbar, part
of the content aggregation of the DynamicPageTitle, is changed, the
GenericTags are getting cut.

Solution: The OverflowToolbar now fires _contentSizeChange event when
the visibility of a GenericTag is changed. The DynamicPageTitle listens
for this event, and if a null parameter is passed (for contentSize) it
resets the CSS flex-basis property to its content container to "auto",
so that the GenericTags can occupy the space they need.

Change-Id: I0cacdb9d0a709426bfb069c1e536f8d660ea5956
BCP: 2070194973