[INTERNAL] sap.ui.integration.Widget: Process translations in the manifest
Solving the following problem:
When widget's manifest contains translations, they were not processed
on widget level and therefore any call to getManifestEntry() inside the
Widget was returning untranslated texts.

By removing the processI18n=false, the translation files are now
loaded 2 times. The second time they are loaded to construct
the I18n model which is used in the component.

Change-Id: I6d1e967dd28e894531709c3f4df327fbc276699e
BCP: 2070248541