[INTERNAL] sap.m.Tokenizer: Refactor _useCollapsedMode method
- This refactoring is part of the InputsAlignment initiative and
targets the _useCollapsedMode of the Tokenizer.
- The method is used to control if the Tokenizer should show/hide
some of its Tokens and the n-more label. However as this is a
private method, it is not ok to be called from the outside.
- The method is needed mainly for switching between two states -
expandend / collapsed (Loose / Narrow). It forces the control to be
expanded by its Tokens or respect its maxWidth.
- The maxWidth property should always be set from the outside, as
there is no proper way for the Tokenizer to know what part of its
parent size it should take.
- With this change, we make use of the new semantic rendering
and we are invalidating the control every time, the state needs
to change (expand / collapse). This way the private method is
called only from within the control and the state is properly
stored in a property.
- Two new properties were introduced in order to expose properly
the functionality to the outside:
  - renderMode - Enum containing two states (for now).
    - Loose - when the Tokenizer is in 'expanded' state and should
      show all of its tokens.
    - Narrow - when the Tokenizer is in 'collapsed' mode and in case
      there is not enough space - it hides some of the tokens and
      displays the n-more label.
  - hiddenTokensCount - The property contains the count of the hidden
    tokens, when in Narrow rendering mode. It is private but has
    public getter. Based on it, the Tokenizer decides if the
    n-more lable will be shown.

JIRA: BGSOFUIRILA-2829
Change-Id: Idcd528b5d581d9958bff63277d4110c9020154af