[INTERNAL] sap.m.WizardProgressNavigator: aria-posinset and aria-setsize attributes added
- Previously, these attributes were not used, because in branching mode
the aria-setsize attribute value was not known.
- According to the ARIA 1.1 standard, 'aria-setsize' property could
receive -1 as a value now. As a result, it could be correctly set both
in regular and branching mode. When the total steps count is not
know, the aria-setsize attribute has -1 as a value. In the other
use cases, its value is the total step count.

JIRA: BGSOFUIRILA-2836
Change-Id: Id314d21921a314d7b358ebba834e4e251f61a12e