[INTERNAL] sap.uxap.ObjectPageLayout: Prevent adjusting header heights
Whenever there is a change in the headerTitle the _adjustHeaderHeights
function is getting called. For example, if the objectTitle or
objectSubtitle are updated, this triggers remeasuring of heights.
However, in some cases this is not needed, as there is no change in the
headerTitle height. This is why, we now do not update cached heights
values, if there is no change in the height of the headerTitle.

Change-Id: I29adf0abdf5003ec09a86c61c1e3f16fa5c6cad1
JIRA: BGSOFUIPIRIN-4606