[FIX] sap.m.OverflowToolbar: Added GenericTags are not cut in DT title
Problem: When the "content" aggregation of OverflowToolbar, placed in
DynamicPageTitle, is empty at first and GenericTags are added later,
they are getting cut, because of the previously set flex-basis property,
which restricts the available space in the title.

Solution: Same as a previous fix (related to changing the "visible"
property of the GenericTags), the OverflowToolbar now fires a
_contentSizeChanged event to the DynamicPageTitle with contentSize=null,
in order to inform it to reset its flex-basis property.

Change-Id: Ifcf3d0a05ef033ba667c405e97d5ce0ed0bdaf83
BCP: 2070257184