[FIX] sap.m.Avatar: Badge icon size adjusted with custom Avatar size
Problem: in Custom size mode, Avatar badge width is a percentage value of the given width
for the Avatar, but the icon size is dependent on the font-size of the parent element.
Solution: We set the custom width as a font-size value of the badge container and the icon
element is calculating its size as a percentage value from the custom size of Avatar for its
font-size.

Change-Id: I15ddcb43ca8e81b196ae51a4d40e67d8b03c7feb
BCP:2080150950
Fixes: https://github.com/SAP/openui5/issues/2925