[FIX] sap.m.ObjectNumber: fixed alignment of the unit part
* until now there was a small vertical disalignment of the number
and unit parts of the control that depended on browser, display
resolution and level of zoom. The problem happens because of different
font sizes of both parts.

* now, after an update of the Visual Design specifications of the
control, number and unit parts have the same size, but use different
fonts, and there is no problem with vertical alignment of both parts
of the control.

Change-Id: I7751b187596315335530c399350a7b0e780c1f5d
BCP: 2070137379