[FEATURE] Demo Kit: Added option for language selection
Now, there is an option for language selection from the information
drop-down panel at the top-right corner.

If the user has chosen to have cookies, the newly selected language
is saved in the cookies. This way, the next time the user visits the
Demo Kit application, the language is the last saved in the cookies.

If the user didn't agree to the cookies policy, a default language is
picked for the Demo Kit. Currently our default language is English.

If there is a language configuration (sap-ui-languge) in the input URI
in the address bar, its value has the highest priority. It overwrites
all of the above configurations.

Change-Id: I9ca723f6d9a6156a5d1429a2837d54a55549d6e1
JIRA: BGSOFUIPIRIN-4627