[INTERNAL] sap.ui.fl: remove support for old selectors without idIsLocal
in the past selectors without idIsLocal (very old changes), whose id
starts with a different appComponent than the current one, would also
work, even though the ID points to a control in a different component.
With the last refactoring this behavior did not work anymore, so we now
dropped the support for this.

Change-Id: Ia44000126c4184e52784526fd9b5884015aff4dd
BCP: 2080131793