[INTERNAL] sap.m.Select: Omitting marking of the Enter event
Problem: The sap.ui.table.Table control, needs to know when the
"Enter" button is pressed in order to create a new row.

Root Cause: That was not possible since we always marked
the "Enter" event. This results in suppressing the "Enter" event
propagation towards the Table.

Solution: Marking of "Enter" event occurs only when the Select's picker
is opened. When there is an opened picker, we mark the event.

Change-Id: I1d0903dfdd46b662ad4e9701ca1ed12be98fb97a
JIRA: BGSOFUIPIRIN-4644