[INTERNAL] m.ListBase: show noData when items are hidden
The ListBase now shows the noData text when all the items are hidden.
It is also necessary to render the noData text to prevent additional tab
stops on ul element and on the dummyAfter element of the ListBase.
In case of growing list, the growing trigger is also hidden when all the
visible items are hidden in the list.
BCP: 2070118563

Change-Id: Iaef67d595eebe7e2dd4620f5cc7fa49043df167d
(cherry picked from commit a2f0795461a123c1ffb2eaf1b14f6f8142878714)
CR-Id: 002075125800002569342021