[FIX] sap.ui.unified.calendar.Month: fixed console error when click on Calendar Card's Header
* originally, this ticket was raized for Calendar Card (CC) - in some cases, when you click
on the Header of the CC, the page is scrolled, and mousedown event happens in CC Header, but
mouseup happens in the Month. There is a method Month.prototype._isValueInThreshold which
checks something (based on mousedown and mouseup coordinates), but expects that if there is
mouseup, to have mousedoun before it. In the case with CC, Month receives mouseup only
(because of the page scroll) and there is no previous this._oMousedownPosition stored,
that's why an error is thrown in the console.

* now this method checks first if there is stored mousedown when mouseup happens.

Change-Id: Ib170c0973a84fefd2ca9e2b2b8863a82b2ad9910
BCP: 2080183678