[INTERNAL] JSDoc: normalize notation for generic types
For generic types, JSDoc allows two different notations, one with a dot
between the type's name and the type parameters (e.g. Array.<string>)
and one without the dot. The documentation for the Google Closure
Compiler (to which JSDoc refers) already rates the notation with the dot
as 'deprecated' (see
https://github.com/google/closure-compiler/wiki/Types-in-the-Closure-Type-System
, Type Application), JSDoc doesn't.

However, as the notation without the dot is also common to TypeScript
developers, all types in the api.json will now be normalized to that
notation.

Change-Id: I88e2e4e3db0a28d9d24be56fc58f5eae98cfeb83