[FIX] sap.m.Breadcrumb: Prevented redundant invalidation
Problem:
When the sap.m.Breadcrumb is rendered for a first time, its internal
sap.m.Select is created and added as private aggregation, which caused
invalidation of both the Breadcrumb and its parent control.

Solution:
Prevent invalidation upon creating the internal select, as this is
happening during rendering and the newly created select will be rendered
anyway (no invalidation needed).

BCP: 2080143818
Change-Id: I9b2e146b17f3ba7c8e14f0a46049a41ee7b25e3a