[FIX] sap.uxap.ObjectPageLayout: Fix updating selectedSection
Problem: When setSelectedSection is called with "null" value and the
first section is already selected, the "selectedSection" association is
not getting udpated (it is still selected in the AnchorBar and the page
is scrolled to it, but getSelectedSection method returns "null",
instead of the selectedSection Id).

Solution: If the ObjectPageLayout is not invalidated in the same time
when the setSelectedSection method is called with "null", we now update
the "selectedSection" association with additional call of
updateSelectionOnScroll. If the ObjectPageLayout is also invalidated at
the same time when the setSelectedSection method is called with "null"
(for example when there is a change in the visibility of its Sections),
we do not call updateSelectionOnScroll because this will wrongly change
the "selectedSection" association. In this case, ffter the rerendering
"selectedSection" is changed by _adjustSelectedSectionByUXRules when
_getSectionsToRender is called by the ObjectPageLayoutRenderer.

Change-Id: I6d1d416528b46c5ec6433301f2308afdb7a2cbe7
BCP: 2080184829