[INTERNAL] sap.m.MultiInput: Added more tests for onsapfocusleave()
Additional check is added in onsapfocusleave method to ensure that
even if no Popover instance is present the function will not throw an
error.

Added two more tests - one is to check that a popover instance is always
present on sapfocusleave and another one that even if by some kind of  mistake in
the future no popover is provided in the function the control will not
throw and break an app.

BCP: 2080177545
Change-Id: Ie2e036c9055ee49ca35dc18e3819b2ed6d1c2ebb