[FIX] "72" font: fix Vietnamese characters with misplaced dot below
Some Vietnamese characters with a dot below, like ạ, had that dot
misplaced to the right in the "72" font.

Those unicode characters are NOT in the Fiori font "72", so some
browsers fall back to the next font in the cascade - Arial/Helvetica,
which works fine.
But some other browsers use a different unicode character (U+323 -
'COMBINING DOT BELOW') to construct the "a" with dot from a regular "a"
and this special combination dot. This special dot is in the "72" font
and was buggy, leading to the wrong location.

The fix replaces the font (only the full version, as the reduced one
does not contain this specific character) with an updated version.

CR-Id: 002075125900001854412020
BCP: 002075129400002309352020
Change-Id: I39d17047e8734b072beac6ae2e3a32fcb91e4241