[FIX] sap.m.MultiComboBox: Input's width calculation is now in decimals
Most of the time the tokenizer's BoundingClientRect width is a decimal
number.	Rounding it to an integer causes visual bugs in some cases
(depending on the tokens' width) and gap appeared between the dropdown
arrow and the input's border.

Now calculation is handled with floating numbers rounded
to 2 decimal points for exact width calculation and cross-browser compatability.

BCP: 2070139347
Fixes: https://github.com/SAP/openui5/issues/2863
Change-Id: I8bade5446912f2660cc5743b587fee5d04808891