[INTERNAL] Table: Fix variable row height scrolling in Tree- & AnalyticalTable
Some TreeBinding adapters and the AnalyticalBinding only update the binding
length when calling Binding#getContexts. When expanding a row in the buffer if
 variable row heights is enabled, this is too late and causes the incorrect
binding contexts to be applied to the rows. Therefore, Tree- & AnalyticalTable
get the contexts from the binding again if the binding length changes after
calling Binding#getContexts and variable row heights are enabled.

Change-Id: I8abf34608fa869072c18dcf19a37f439ede717d7
BCP: 2080135440