[INTERNAL] sap.ui.integration.widgets.AdaptiveContent: OpenURL action sample added
The first adaptive card sample is updated to use sap.f.cards.Header.
Header is also added to the Submit Action sample.
New Action.OpenURL sample is created within the Adaptive section which is now also included as a
link in the documentation page.

New sample is created in the Integration section - Adaptive With Customized
Actions. It shows how the actions implemented in the adaptive cards can
ibe modified with custom handlers in the application.

JIRA: BGSOFUIRILA-2841
Change-Id: Ib8b4bc6c57aa1f096ec2cedc0372d66846db5a13