[FIX] sap.m.TimePicker: on focus in behavior over the input is now consistent
Problem description:
When on change event is fired over the input field of the TimePicker control,
then the value could be adjusted internally.
This is unexpected behavior for the sap.m.MaskEnabler temporary values.

Solution:
Now the sap.m.MaskEnabler temporary values are upadated after there change
event is fired in the sap.m.TimePicker.

Change-Id: I55fd4774fc7ebf92fb2adba33ba2942b7dcf7cb0
BCP: 002075129500003307462020