[INTERNAL] sap.m/f.SemanticToggleButton: Use ToggleButton correctly
Problem: SemanticToggleButton in sap.f and sap.m libribries are not using
sap.m.ToggleButton and instead are changing button type from regular to
emphasized, which leads to visual problems in HCB/HCW themes. Buttons
are hardly distinguishable.

Solution: Use ToggleButtons as they have a clearly visible toggle selected
state in all themes and are not the same as an emphasied button.

Change-Id: If4fdd3bc535958a1e2a5946db41920294282ab8b
BCP: 1980289056