[INTERNAL] sap.ui.unified.Menu: visual tests now run successfully
Problem description:
A visual test was checking if the focus and hover remained on
the last focused sap.ui.unified.MenuTextFieldItem, when the mouse
is moved out of the sap.ui.unified.Menu instance.
This scenario is no longer valid, as the hover state doesn't remain
on mouse out of the menu. Also the focus is indicated visually with
blinking a caret symbol inside the input field, which causes the
test to be unstable.

This change reverts the gerrit change with id: 4726145

Change-Id: I86ff982fc5d2f25579798fd66e89aeaed827a31e