[FIX] sap.f.DynamicPage: Prevented unwanted scroll
Problem:
When the header is in the overflow of the scroll container,
it still takes space and whenever its height changes,
it affects the scroll position of the scrollable content bellow it.
Whenever its height increases/decreases, the content bellow it is
pushed down or up, respectively => visual jump of content

Solution:
Whenever the header height changes, we adjust the scroll position
accordingly, to compensate the increase/decrease of header height.

Change-Id: I51681e1145753d423582278516112b019c804786
BCP: 002075129400002536912020