[FIX] sap.uxap.ObjectPageLayout: Fixed background color of Spacer
Problem:
The background color of the ObjectPageLayout's Spacer is missing.

Root Cause:
A CSS rule states that the background color should be applied to the
immediate DIV (being the Spacer) element after the ObjectPageLayout's
Section container.
A recent regression broke the CSS rule by putting a SPAN to be the
immediate element after the ObjectPageLayout's Section container.

Solution:
We swap the places of the SPAN and the DIV in order to keep the CSS
rule in working condition.

Change-Id: I279de34067ba352f8194c6f2bfbf4a007df4e3cd
BCP: 2080156739