[INTERNAL] Demo Kit: Remove new visual design of technical entities
Problem: A new visual design for technical entities was introduced, but it does
not look good in some cases (ex. in table heading) where the backgrounds
are similar and is also not consistent with the technical entities from
API Reference tab.

Solution: Return the old styles of technical entities

Change-Id: I70f03c98dcf849445c8ca4e0b9aa5da566e1139a
JIRA: BGSOFUIPIRIN-4688