[FIX] sap.ui.unified.FileUploader: upload input element is now created only once
Problem description:
When using mime type or file type setters after the control is rendered,
then two upload input elements were created.
The first upload input element was appended to the sap.ui.unified.FileUploader markup.
The second upload input element was appended to the static area and it was used
in order to track the selected files.

Change-Id: Id07dcb6eedeb3786a20622ce1d555a744a2ec6a0
BCP: 2070211785