[FIX] sap.f.DynamicPage: Fixed hiding of footer when animations are disabled
Problem: Footer can't be hidden when the animations of the framework
are disabled.

Root Cause: We were expecting a certain configuration of the framework
regarding the animations to be 'none', but when the animations are
disabled with 'data-sap-ui-animation="false"', the configuration is
'minimal'. We were trying to run an animation when we are not supposed to.

Solution: We now check if the animations configuration of the framework
is whether 'none' or 'minimal'. This fixes the 'hiding' bug, since
no animation takes place.

Change-Id: I095e330c531dcaa8606ee593ba822661922a4fff
BCP: 2080206356