[INTERNAL] Demo Kit: Fixed incorrect focus of the application's logo
Problem: The focus of the main logo of the Demo Kit is incorrect.

Root Cause: The logo is inside html anchor control, therefore the focus
is applied to the anchor and not the image.

Solution: We remove the anchor and leave only the image. The image is
sap.m.Image therefore inherits the correct css styling for the focus.

Change-Id: If7d21b64743a4888b6086d1e199537c7baff5ac7
BCP: 2080218935