[FIX] sap.m.BadgeEnabler: Redundant show of empty badge fixed
Problem: When control has been initially rendered and badgeCustomData has been attached to it,
but without a value, an empty badge was shown, because of an invalid check of its existence
Also truncation was corrupted.

Solution: We now check not only of the data attached, but of the validity of its value.
For the truncation to work we put "nowrap" value to white-space property.

Change-Id: I832eadd076357dd3f073c9ae2f9e66e86a356d2e
JIRA: BGSOFUIPIRIN-4690