[INTERNAL] m.Table: add dummy column when all columns have a width defined
The Table control adds dummy column when all the columns have a width
defined. If the dummy column is rendered by the table and growing button
is also available, then the growing button width is also adapted to the
sum of visible columns width.
Minor bug fixes for autoPopinMode.
Adapting GroupHeaderListItemRenderer with RenderManager apiVersion=2
Minor bug fixes to GroupHeaderListItem as well where the NavigatedCol
was not rendered for the GroupHeaderListItem. This lead to bad table layout.
The NavigatedCol is not added to the GroupHeaderListItem to ensure table
layout stability.

JIRA: CPOUIFTEAMB-1556
JIRA: CPOUIFTEAMB-1595

Change-Id: I01556631ed65f2e9ad635794912ba1e080c1b748