[INTERNAL] Demo Kit: Corrected table width and column width
Columns should be proportional if the colwidth attribute is set.
For example, if each column (in the <colspec> element)
has colwidth = "1*" it means we should aim to display them proportionally
if there is enough space available.

If the colwidth is different for each column, we should aim to
display as it is set, for example, the first column can be 2*,
second can be 1* and third can be 3* which means the first
should have double the width of the second column and the
third column should have 3 times the width of the second column.

Additionally, if the pgwide="1" attribute is set in
the <table> tag, the table width should be scaled
to take all the available width even if the content is
not enough to require such width.

Change-Id: Ia2f5e483d622a2bd418427c716cccdab199109a9
JIRA: BGSOFUIPIRIN-4680