[INTERNAL] AnalyticalTable: Fix level announcement
If a total summary row is visible, the levels of all other rows are increased
by one. Therefore, this row has level=1, whereas other levels start at 2.
This makes the screen reader announce the level when moving from/to this
row, even if the row above is visually on the same level.

The levels are like this mainly for technical reasons (the base is the level the
table gets from the binding). This is changed to reflect the visual state.

Change-Id: I42ba509238c36519e06ac218e768499545ebb5ce
BCP: 2080176774