[INTERNAL] sap.uxap.ObjectPageLayout: Revise spacer logic
Problem: When there is a sap.m.Table, placed in a single
ObjectPageSubSection, which takes the whole available space, spacer's
adjusting logic of the ObjectPageLayout leads to a problem with showing
"Show More" button of the Table.

Solution: The spacer is set to "0px" height when there is a single
SubSection.

Change-Id: I7396415f45dc6dec48c13a1eb19f78c2f54c605a
JIRA: BGSOFUIPIRIN-4295