[FIX] sap.m.Button: drag and drop now works on safari and firefox
Firefox and Safari doesn't set the focus on the clicked button, but
on the element above it, so previously we've had to prevent the default
behaviour and set the focus manually. The prevention of the default behaviour
of the browser, we lost the Dnd functionality. With timeout we're moving the
focus function on the end of the event loop so it passes after the browsers
focus function and we don't have to use preventDefault. We have tests for the
focus on both browsers and the dnd event fire was tested manually.

BCP: 2080226059

Change-Id: Ib87749675926d8800198f19521ce8edf42f57717