[FIX] sap.ui.unified.Month: Dummy cell's visibility fixed
On accessing dates within the calendar, the incorrect week day name
was associated and read out instead of the currently accessed one.
To fix this we removed the dummy cell's hidden visibility, as it
was affecting screen readers.

One additional if is also downported, as in PlanningCalendar's
OneMonth view those week numbers aren't needed at all.

Change-Id: I82f761a8bf632f30e91df9dd6b52458a38b282cd
CR-Id: 002075125900002479332020
BCP: 2070244746
BCP: 2080217137