[FIX] FieldValueHelp: unneeded get call to determine item
The FieldValueHelp is assigned to a Field showing the description.
User enter a valid key, the suggest list shows this key (and maybe
more).
After ENTER or Focusout the FieldValueHelp is used to determine the item
for the entered value.
At first it looks if it is a valid description. For the entered value
(as it is a key) no item with this descriprion is found. A filter
request is triggered.
After this filter request don't find a item, the FieldValueHelp checks
it it is a valid key. This is found in the already known items and
returned. 
So the filter request for description is not needed.

Change-Id: Ia819336a94cf38c4aab201cad1e764564f9a95d0
BCP: 2080208502
JIRA: CPOUIFDENVER-1477