[FIX] sap.m.OverflowToolbar: Ensure extra space in DynamicPageTitle
Problem: In some rare cases, the width of the OverflowToolbar is
measured as 165px (when it is 165.994px), and as it was previously
166px, one of its Buttons go in the Popover, when there is enough space
for it in the DynamicPageTitle.

Solution: Overcome rounding issues with allocating 1px extra for the
OverflowToolbar in the DynamicPageTitle, which sets its flex-basis at
the extact width value, enough for the OverflowToolbar's content.

Change-Id: Ibdb96ecc9307707e7dca741bc5bc55c259527417
BCP: 2080160972