[INTERNAL] JSDoc Template: handling of defaultValues, error logging
A default value of "" (empty string) was not handled properly when
generating auto-doc comments and a default value of "undefined" was not
represented in the final api.json.

To avoid confusion in the log, errors that do not break the build are no
longer reported as errors, but as 'future errors'.

APIs that can only be addressed via their global name, not via an AMD
dependency, can now be marked with @ui5-global-only. They are no longer
reported as errors during the build (as they can'T be fixed).

Change-Id: Ia2c9a4686a5901cf590c7957f0f74894641855ac