[INTERNAL] sap.m.Tokenizer, sap.m.MultiInput: Refactor validation logic
The validation logic of the sap.m.Tokenizer makes sense
only when used inside input controls. The tokenizer purpose
should be to store tokens and act as a container for them.

With this change we removed the validation logic from the Tokenizer
and added it to the MultiInput control.
All methods are revised, refactored and simplified. The public methods are deprecated from the Tokenizer and documented. All of the methods, which
are changed are described below.

Deprecated from Tokenizer:
addValidator
removeValidator
removeAllValidators
addValidateToken

Removed from Tokenizer:
_tokenExists - moved to MultiInput
_addUniqueToken - moved to MultiInput
_addValidateToken - removed, since it is the similar to addValidateToken public method
_getAsyncValidationCallback - moved to MultiInput
_validateToken - moved to MultiInput
Tokenizer.WaitForAsyncValidation - MultiInput.WaitForAsyncValidation should be used instead

Newly created methods in MultiInput:
getValidators - public
_configureTokenOptions - private

JIRA: BGSOFUIRILA-2813
Change-Id: Idfb200c6954e8ec039d27477a1c578235a594452