[INTERNAL] duplicating the sap.ui.fl...FileListBaseConnector
With higher UI5 versions the
sap.ui.fl.initial.api.connectors.FileListBaseConnector was moved (before
active consumption) to the write package.
This is needed to have a clear separation on namespace level of use
cases (designtime & previews vs. runtime).
To avoid forcing consumers to introduce a version switch the "move" is
down-ported to 1.78.
This commit only adds the file without removing the original file (which
is not available in higher versions) to minimize the impact on the 1.78
code line.

Change-Id: Icedf95fe655461792d4f230543745f9b4d8f64ca
BCP: N/A
CR-Id: 002075125900002572872020
Signed-off-by: Christian Voshage <christian.voshage@sap.com>