[INTERNAL] sap.ui.integration.card.AdaptiveContent: Avoiding redundant data requests
Previously there were cases that the request to a data service such as
Northwind was duplicated, now the behaviour is corrected and only one call is executed.

The issue with the extra data request after card invalidation is also
fixed.

Creation and configuration of the DataProvider is refactored and delegated
to the BaseContent.

New "card" association property in the metadata of BaseContent is added.
It is set by the ContentFactory.

Loading state checking in BaseContent is adjusted to take into account
case when the card is not yet created.

LoadingProvider is implemented for managing the loading state. Loading
Placeholders (temporary content) for the content that is still not fetched is to be added in
the future instead of "busy" state.

QUnit tests are modified and extended.

JIRA: BGSOFUIRILA-2877
Change-Id: I2ed8cbe32de717b6c1bf5575412081c028dc60cd