[INTERNAL] sap.m.PlanningCalendar: A Demokit sample with pagination of rows
* Because of poor performance of the control when there are many rows with
many appointments per row, there is a need of pagination of the rows in order
to display only few of them at once.

* this change presents a functionality outside of the control, that enables
pagination of the rows. There is a configuration that allows to the developer
to use fixed or user-selectable number of rows per page, and also page change
method - by dropdown or by buttons

Change-Id: I72fd1cdac1d45c7fd9ade8f3c742ca98ed559b7f
JIRA: BGSOFUIBALKAN-5418