[INTERNAL] sap.f.GridContainer: Fix drop indicator position for IE
Solves the following problem:
When dragging over a grid in IE, the drop position
was shown not on the target, but on the next item.

This was caused by two dummy areas in the grid needed for the focus handling.

The problem appeared on Chrome as well.
There the dragged item had to go way after the drop item
in order to show the indicator, which is not desired behaviour.

Change-Id: I1d912682fa54b01a304a15ffe179a9fa4b2764f3
BCP: 2080187242