[FEATURE] ListFieldHelp: sap.m.Select-like behavior enabled
For Select-Like controls a ListFieldHelp bound to a static set of items
should be used.

This commit includes the following changes:
- ListFieldHelp: new property filterList to prevent filtering on
type-ahead. So all items are displayed while typing.
- ListFieldHelp: new property useFirstMatch. So if only the starting
part of the item text is entered into the Field, the first matching item
will be used on ENTER or focusout. On typing the first match will be
shown as selected in the list.
- FieldHelpBase: new function openByClick. Disabled by default. If in
ListFieldHelp filtering is disabled, the list opens if the user clicks
into the field.

Change-Id: I908e4f45882bb47d9a3fce0c8520d41230536125
JIRA: CPOUIFDENVER-1443