[INTERNAL][FIX]ColumnLayout: use translateX for Safari
As the ColumnLayout will use column-count for the wrapping <div> element
to display different amounts of columns, the overflow:hidden on the inner
FormElement <div> causes an issue to only display the items within the first
column as visiible causing containers with a column-count higher than one in
edit mode to not display the input fields as expected (even though the elements are corretly positioned in the DOM with the according styling).

Using transform: translate does trigger a recalculation for the positioning of the inner FormElements <div> containers causing the Input fields to show up as expected. As this is a workaround for Safari, the styling will only be added for Safari browsers.

Change-Id: If5e5ebabb91c2ecc92cada9fbaa9242bdd42d5bc